Create Optimus Array Question
Given two integers N and M as an input. N denotes the length of circularly connected array and M is an independent integer. Starting from first index of array we need to fill integers on indices as 1, 2, 3 and so on such that the sum of integers filled doesn’t exceed M….. For example:
N = 3, M = 6. Then array = {1, 2, 3}
N = 5, M = 8. Then array = {1, 2, 3, 2, 0}. 3rd index is filled with 2 at place of 4 because if we put 4 there, then the sum of filled integers will exceed M = 8.